⚡️ Use data `vault_namespace` to ensure namespace exists
As namespace is not deployed with this module, if the `var.namespace` is
defined, this could lead to issues while trying to a non existent namespace or
to a namespace not permitted to identity associated to the credentials used with
th provider.

This commit update the submodule to ensure either the namespace exists or access
is permitted through the usage of data.